Подготовил решение KTB-21

+++ app/src/main/kotlin/TelegramBotService.kt
@@ -17,7 +17,7 @@ class TelegramBotService(private val botToken: String) {
     private val client: HttpClient = HttpClient.newBuilder().build()
     private val json = Json { ignoreUnknownKeys = true }
 
-    fun getUpdates(updateId: Int): String {
+    fun getUpdates(updateId: Long): String {
         val urlGetUpdates: String = TELEGRAM_BASE_URL + botToken + "/getUpdates?offset=$updateId"
         val request1: HttpRequest = HttpRequest.newBuilder().uri(URI.create(urlGetUpdates)).build()
         val response1: HttpResponse<String> = client.send(request1, HttpResponse.BodyHandlers.ofString())
